+++ CHANGELOG.md
@@ -13,7 +13,9 @@
 - Genesis state is now configurable during bootstrapping (#1000)
 - Added configurable network id for the faucet (#1016).
 - Network transaction builder now tracks inflight txs instead of only committed ones (#1051).
-- Add Otel trace layers to `miden-remote-prover` and `miden-remote-prover-proxy` (#1061).
+- Add open-telemetry trace layers to `miden-remote-prover` and `miden-remote-prover-proxy` (#1061).
+- Add open-telemetry stats for the mempool (#1073).
+- Add open-telemetry stats for the network transaction builder state (#1073).
 
 ### Changes